requirements.txt
playwright==1.42.0
pytest==8.1.1
allure-pytest==2.13.2
requests==2.32.0
python-dotenv==1.0.0
pytest-xdist==3.5.0